html anchor in head (#5071)
