+++ test/runtime/samples/raw-mustache-inside-slot/Component.svelte
